requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
openai-whisper
numpy
ffmpeg-python